pyproject.toml
[project]
name = "promoagent"
version = "0.1.0"
description = "Autonomous Reddit Marketing Agent"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "praw>=7.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "loguru>=0.7.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
